requirements.txt
pandas
speedtest-cli